refactor(article-list-card): split file in different components
